requirements.txt
# Core deep learning
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0

# Experiment tracking and logging
wandb>=0.13.0
tensorboard>=2.10.0

# Data science and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
pandas>=1.4.0
scikit-learn>=1.1.0

# Image processing and metrics
Pillow>=9.0.0
opencv-python>=4.6.0
scikit-image>=0.19.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.64.0
hydra-core>=1.2.0

# Interactive demos
streamlit>=1.12.0
gradio>=3.4.0

# Development and testing
jupyter>=1.0.0
ipykernel>=6.15.0
pytest>=7.1.0
black>=22.6.0
flake8>=5.0.0

# Optional: For advanced metrics
lpips>=0.1.4